pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sec-xbrl-finwarehouse"
version = "0.1.0"
description = "SEC XBRL company facts -> Postgres -> FastAPI (ratios + screener)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "psycopg2-binary>=2.9.9",
  "fastapi>=0.110",
  "uvicorn>=0.27"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]